+++ src/Client.php
@@ -22,6 +22,7 @@ use Gitlab\Api\Groups;
 use Gitlab\Api\GroupsBoards;
 use Gitlab\Api\GroupsEpics;
 use Gitlab\Api\GroupsMilestones;
+use Gitlab\Api\Integrations;
 use Gitlab\Api\IssueBoards;
 use Gitlab\Api\IssueLinks;
 use Gitlab\Api\Issues;
@@ -170,6 +171,11 @@ class Client
     public function groupsMilestones(): GroupsMilestones
     {
         return new GroupsMilestones($this);
+    }
+
+    public function integrations(): Integrations
+    {
+        return new Integrations($this);
     }
 
     public function issueBoards(): IssueBoards