+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add support for `symfony/options-resolver:^8.0`
 * Add support for `visibility` in `Groups::all`
 * Add support for personal access tokens
+* Add support for project integrations endpoints
 * Add support for `job_inputs` and `job_variables_attributes` in `Jobs::play`
 * Add support for filters in `Projects::projectAccessTokens`
 * Add support for `Projects::rotateProjectAccessToken`